% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_col.R
\docType{data}
\name{nu_col}
\alias{nu_col}
\title{NU Color Function}
\format{An object of class \code{character} of length 15.}
\usage{
nu_col
}
\description{
This function pulls in Northwestern colors
}
\examples{
nu_col
nu_bright
nu_dark
}
\keyword{Northwestern}
